Use is_intrange() to validate QinQ tag range
Note: is_intrange() only tells us that the range is good and might be delimited by a "-" or a ":", so if we want to take apart the 2 parts of the range (like is done here) we have to also test the string to decide which separator to use.
I decided it is fine to accept ranges with a ":" separator. Even though it is not explicitly mention on the UI, a user who accidentally does it will be pleasantly happy that it works.