system_advanced_admin remove href from setHelp strings
I think that giving long strings that contain various HTML tags in-line through for translation is going to be error-prone. Actually we do not want the translators to accidentally also translate href like system_usermanager or accidentally miss out some quotes or brackets. It seems safer to substitute these in using %s formatting.

This is an example of the sort of changes that could be made.