Update misleading help text
Commit d57725aac5145b4f17097d61e3b5a6ca72f1754e updated the help text
to use LDAP search filter syntax. This is misleading, however, as
the auth.inc file already sets the LDAP filter syntax for you based on
the input string in this field. This field should be the bare qualified
search string instead, and using the full syntax could be confusing to
end users.