Mount the disk (the root slice) as read-only before running the fsck in the preen mode.
This force the flush of all pending writes to disk and thus put the disk in a consistent state.

The second effect of this change is that the file system can actually be recovered now, because otherwise with the root slice mounted as read-write, the fsck cannot write any changes.