Unbound: Disable IPv6 outgoing queries if IPv6 blocked in firewall, as they can never go anywhere
If IPv6 is disallowed in system->advanced->network, then any IPv6 lookups by Unbound will always be blocked, so there's no point sending them.

The practical purpose is that they also clog up the log and may fractionally slow down the resolver because the resolver then has to deal with IPv6 not replying, fallback lookups, etc.

I'm aware that we state that the system setting doesn't affect any service, but in this case the service simply doesn't send packets that we know will be blocked anyway on egress, so it's probably not going to be an issue. If it is, we can add an explicit "Disable IPv6" setting to this and other functions, but then if IPv6 is later used, the user has to remember to unset "IPv6 not allowed" in all the pages they were set, which is a pain and pretty much duplicates the main setting for all practical purposes.