"Save" user selection for DUID type
It proved confusing for users who entered a DUID as a certain type to
see the resulting DUID file displayed as a "Raw DUID" upon page reload.
Still work with dhcp6c_duid file data, but parse it and show the Form
corresponding to the DUID type with the inputs filled in, so the user
sees what she previously entered.

Raw edits to the DUID file are still possible.