Fix #8290
On d9b05eb490a the way aliases containing a mix of IP address and FQDNs
works has changed and all items were added to filterdns to make sure
the pf table created would be persistent. The flag $use_filterdns was
being set only when the array item used to control didn't exist yet.
Because of that, the second call to the same function made it to return
a different result. Set it independent if a new array item is created
or not to make the function to properly return the right thing.