Allow Dual Stack IPsec P1 interface. Fixes #6886
Allow "Both" to be selected for IPv4/IPv6 on IPsec P1, in the config use both addresses as "left =" if they both exist.
Some cases where a single address was assumed (e.g. ping hosts) default to using the first address.

Doesn't make sense when selecting a VIP that is a single type, but should still handle that gracefully, only using the one that actually exists.

There might be some edge cases that this doesn't handle properly.