Add array check
Even though I now set `$ns` equal to `array_unique(get_nameservers()`, just to be safe we check with `is_array($ns)` and set it to a blank array if not. Shouldn't ever happen, but an extra error avoidance test is better than a hard to track down bug later.