on arm and arm64 machines, set kern.shutdown.secure_halt = 1
the arm systems leave enough running after halt to forward packets.
this is a bad thing. on arm systems, set this sysctl so that when
a halt command is issued, it is severely stopped and no packet
forwarding happens. note that this also breaks the ability to reboot
from halt using a keypress (and the halt messages say so).