Make the WF2Q+ the default scheduler for the dummynet limiters.
The WF2Q+ was the default scheduler in previous versions, it is well tested and support dynamic queues.

Add a note for the FIFO scheduler to make clear that it does not support dynamic queues (by design) and as such, it is working as intended.

Add the scheduler information to Diagnostics -> Limiter Info.

Ticket #8973