Certificate date calculation changes. Fixes #9899
Make the certificate date calculation more general and also try multiple ways
to determine the date (both timestamp and unix timestamp).

Catch cases where one or the other date fails to calculate to avoid errors.

While here, also account for the fact that DateTime::diff result is always
positive, when we need to know if it's negative.