Fix #9612: Reboot twice if needed and run fsck -z
After changing pfSense-rc to create /.fix_for_SA-19-10.ufs file to flag
system to run fsck -z during pre-install stage it started failing
because fsck binary, at that point while only kernel was upgraded, still
don't have -z flag implemented.

This changes the way rc script deals with it and check fsck return code.
When it fails (!= 0) flag is not removed and later, after all base
packages are upgraded, call a reroot (if possible), or a reboot and then
the same flag is used and fsck -z is triggered.

To make it easier to users to identify if this fix was applied, it
creates the file /cf/conf/applied_fix_for_SA-19-10.ufs when it succeeds.