RAM Disk robustness improvements. Fixes NG 3173
* Prevents RAM disk from being partially enabled and left in a broken
state if the RAM disks cannot be created
* Prevents RAM disks from potentially overcommitting free kernel memory,
which FreeBSD now prevents, but could result in errors.
* Adds GUI input validation to prevent exceeding free kernel memory when
sizing RAM disks
* Ensures the pkg database is always available whether setup for disk
storage, RAM disks, or if set for RAM disks but they failed to work
* Moves RAM disk script functions to a common file for reuse.
* Files a notice to alert the user that RAM disks failed and to check
the disk sizes
* Generally makes the RAM disk setup process more verbose on the
console. These messages can be commented out or removed in the future,
but are helpful for the time being.