Add help text to Register DHCP Leases in Resolver
This change adds a sentence to the help text for the `Register DHCP Leases in the DNS Resolver` input field. 

Enabling this option has the non-obvious side effect of reloading Unbound each time a client obtains a DHCP lease. This reload flushes the resolution cache, which can mean long recursive lookups being required for even frequently used domains. 

A long term fix for this issue would be to modify the Unbound scripts so that the resolution cache is persisted somehow between reloads, but in the short term this addition should help users avoid this pitfall.

This issue comes up in the forums from time to time, and I know that personally I long avoided using Unbound in recursive mode because I couldn't figure out why performance was sometimes poor on frequently used domains. Turns out it was because I had this option turned on.

Some forum mentions:
https://forum.netgate.com/topic/146117/dns-resolver-not-caching-correct/56
https://forum.netgate.com/topic/154368/unbound-resolver-low-cache-hits