build: Fix pkg.txz.sig link
When pkg transitioned from .txz to .pkg extension we enabled an option
to keep a symlink of old extension around and make it possible for users
running older versions to upgrade.  At that time, when we sign bootstrap
pkg, we changed scripts to create .sig symlink as well.

The default is .pkg for a long time, invert the logic and consider
pkg.pkg.sig as main file and create a link called pkg.txz.sig.