OpenVPN shared key warning. Implements #12981.
Adds a warning to the OpenVPN client and server list and edit pages
warning the user about shared key mode being deprecated by OpenVPN.

Warning only displays on the instance lists if there is an existing
shared key instance. Warning only displays when editing an instance
if the instance is set to shared key.