ovpn-dnslinkup: do not set interface routes for DNS servers
If the OpenVPN server provides DNS server information (and 'Pull DNS' is
set) we add these DNS servers to our DNS configuration.
That's fine, but we should not add interface routes for these DNS
servers, because there's no guarantee they are on the 'local' VPN
network.
We should also not remove such routes if we never create then in the
first place, so also update ovpn-linkdown.

One example is the Netgate OpenVPN server, which provides DNS servers
172.27.32.5, 172.27.32.6 and 162.208.119.38. These do not live in the
OpenVPN subnet (172.27.114.128/25), so we cannot reach them directly.
Note that this does actually work with OpenVPN in non-DCO mode, but
that's an implementation artefact of OpenVPN, not behaviour we can rely
upon. (As demonstrated by the fact that it does not work with DCO
enabled).