Add config_get_path and config_set_path for config element access
config_get_path and config_set_path allow the user to specify a configuration
element by path rather than by direct access to the global config array. This
has several advantages, including eliminating errors attempting to access
subarrays that may not exist in a configuration and abstracting the actual
implementation of the configuration structure from the codebase so that the code
is not dependent upon it.

A small set of phpunit tests are supplied, although there is no automated
mechanism for running them at this time.