CA/Cert descr validation fixes. Fixes #13387
Validate description on save when editing and in other situations that
were not yet covered.

While here, ensure that errors when editing a cert leave the user on the
cert edit screen properly, but successful cases return to the cert list.

Also encode some output just in case a bad value was already present
before the validation was fixed.