Fix text_to_compressed_ipv6() to omit %ifname from v6 addresses
Prior to this, v6 addresses would be considered by Net_IPv6::compress() to have
an ipv4 part consisting of all text from the address from the final colon
separator, and try to treat it as such. This is now problematic in php 8 as when
this compressed address is checked for what type of address it is an invalid
arithmetic expression that was silently ignored in php 7 now generates an
error. This is corrected by trimming the %ifname text from the address prior to
sending it to Net_IPv6::compress() from our text_to_compressed_ip6() utility
function.