Fix services installation only installing the last service of a pkg. Issue #13446.
The original  re-fetched the services list from the config at every iteration
over the package's services list when adding services during installation. After
moving to a get/set model for the config, this discarded service added at each
iteration and resulted in only the last service in the package's list being
added to the config. This is fixed by moving the services fetch prior to the
loop.