Fix array_get_path() not returning $default for null-like values. #13446
In array_get_path(), a $default provided that is not null is intended to invoke
alternate behavior where if the path exists and is empty, $default should be
returned. This requires not identical compare as opposed to a not equal compare,
as some empty values such as [] are considered equal to null.