Replace direct config accesses in auth.inc. Issue #13446
Additionally, change local_group_del_user() to include the index of the group in
the path to the group's member list when removing the user from it. This appears
to have been broken in the original code. With this fixed, remove the redundant
call to local_user_set_groups() in local_user_del that has the same effect and
was likely added because local_group_del_user() was not working correctly.