Correct codelq shaper input validation for firewall_shaper.php. Fixes #13661
Ensure all bandwidth values are cast to int before applying arithmetic to the
return value of get_bandwidth_typescale(). This alleviates failed validation
when the bandwidth is blank.

Two more validation errors are also corrected by only validating qlimit and
tbrsize are positive integers if their input values are not empty strings. These
are also modified to check for values < 1 rather than < 0 to agree with the
validation that the value is positive, which does not include 0.