captiveportal: actually allocate a pipe number for new clients
When a client authenticates to the captive portal we generate a pipe
number (actually two) for it. However, we did this with
'check_only = true', so the next client got assigned the same pipe
number. This resulted in all clients being disconnected (because their
traffic got sent into a pipe which no longer exists) when one client was
disconnected.

We no longer pass the result of the check_only
captiveportal_get_next_dn_ruleno() to portal_allow() so that we do
actually assign unique pipe numbers for each client.

While we're here we also remove the incorrect
captiveportal_free_dn_ruleno(), as the $pipeno here isn't allocated, so
also doesn't need to be freed.

This builds on the logic of #13204, in that we still check if there are
available pipe numbers without allocating one until we've actually
authenticated the user.

Redmine:	#13488