Make ArrayGetExprRector configurable for multiple variables
Rector only permits a single instance of a Rule class in a configuration, so in
order to allow the ArrayGetExprRector rule to be used for more than a single
variable its configuration must accept multiple sets of var/accessor pairs. The
test is additionally restructured to have a common config and to select the
fixtures based upon the test name.