Remove invalid syslog test. Fixes #14283
Having an empty syslog config section is valid, as is not having a
section at all. We should always generate the configuration since we
always start the syslog daemon.

If we want to eventually allow not configuring or running the daemon at
all, we should add a dedicated option for that, but as-is this test
wasn't doing what it implied.