Stop sshguard spam in system logs. Fix #12747
Instead of restarting sshguard with each log rotation, keep it running
separately while monitoring the auth log file. This allows sshguard
start/stop events to still be included in the system log without
spamming it. When the log file is rotated by newsyslog, sshguard will
handle it and hence does not need to be signaled.

The use of deamon(8) allows the service to be restarted automatically
(every 60s) unless stopped via the service control mechanism used by
the GUI. It also prevents the start of multiple sshguard instances.