Reduce rc.newwanipv6 actions on RENEW
This change passes the dhcp6c REASON to the script. The script uses this
to retain the previous behavior for REASONs other than RENEW. If the
reason is RENEW, action is only taken when there has been a change. Also
include the REASON in the info log message.

Followup to a008e2a0865932096129915fcd847df6272ffc96.