Respect the RR type implied by the DDNS Service Type. Fix #16045
Background:
In previous pfSense versions, the RR type and the IP address family used
to make the DDNS request were assumed to be the same and could be
overridden for specific DDNS services. The implementation of #11177 avoids
this assumption in order to support updating the RR over IPv4 and IPv6
regardless of the RR type. This change introduced potential issues when
the configured "interface to monitor" contains both IPv4 and IPv6
addresses:
- The DDNS request may be sent using an IP address family that is not
  supported by the DDNS server / API endpoint.
- The RR type defaults to the request's IP address family which may not be
  the same as what's intended by the DDNS Service Type.

Commit:
This commit addresses the second issue by separating the RR type from the
IP address family used to make the DDNS request. The RR type is once again
based on the DDNS Service Type.

Note:
The first issue may be resolved by creating a gateway group with a single
gateway and set as the "interface to monitor". Alternatively the code for
the relevant service can be updated build the cURL request using the
address family supported by the server/endpoint.