Fix serialization/deserialization of ppp hostuniq and provider attributes
When these two properties were added, the form handling set their config values
to boolean true if empty, leading to empty tags in the config.xml which would be
serialized to the config.cache as empty strings. As a result, later fetches of a
ppp config from the config after the cache is reloaded (i.e. additional
requests) deserialize this true value as an empty string.

However, with the pfnet-controller service running and acting as the config
provider, the raw array written to the config is directly serialized which
would result in hostuniq and provider being stored as and returned as boolean
true rather than the expected empty string, causing a bogus hostuniq and
provider value to be written to the mpd configuration (or passed on to the
if_pppoe ifconfig, as applicable).

To remedy the situation and retain compatibility with config caches that have
the boolean values, interface_pppoe_get_hostuniq(), interface_pppoe_configure(),
and interface_ppps_configure() are changed to check that these attributes are
nonempty as well as string type before attempting to use them. interfaces.php is
modified to not set these properties to true when not supplied in the form.

(cherry picked from commit ff0af2d353b8db1eba524371556d51d028f03dae)