get_interface_addresses(): Change use of 'tunnel' for p2p destination addresses to 'dstaddr'
In the php module, the 'tunnel' address attribute returned from
pfSense_get_ifaddrs() has been renamed to 'dstaddr' to more appropriately
reflect that the address is a p2p destination. This change alters the behavior
of get_interface_addresses() to reflect that difference both in consumption of
of the address property and presenting p2p endpoint addresses as interface
attributes as dstaddr and dstaddr6, appropriately. The only consumer of
this, the pppoe_handler script, is adapted to the change as well.